Fixed general helpers fase 1
