+++ cloudkitchen/templates/base/nav.html
@@ -12,7 +12,7 @@
     <li class="nav-item dropdown" id="link-sales">
         <a class="nav-link dropdown-toggle" href="#" id="supportedContentDropdown" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Ventas</a>
         <div class="dropdown-menu" aria-labelledby="supportedContentDropdown">
-          <a class="dropdown-item" href="{% url 'sales:new_sale' %}">POS</a>
+          <a class="dropdown-item" href="{% url 'sales:sales_new' %}">POS</a>
           <a class="dropdown-item" href="{% url 'sales:sales' %}">Analytics</a>
         </div>
       </li>
