+++ helpers_origin.py
@@ -12,99 +12,6 @@ from products.models import Supply, Cartridge, PackageCartridge, CartridgeRecipe
 from sales.models import TicketBase, TicketDetail, TicketExtraIngredient
 
 
-class Helper(object):
-    def __init__(self):
-        self.tz = pytz.timezone('America/Mexico_City')
-        self.days_dict = {
-            'MONDAY': 'Lunes',
-            'TUESDAY': 'Martes',
-            'WEDNESDAY': 'Miércoles',
-            'THURSDAY': 'Jueves',
-            'FRIDAY': 'Viernes',
-            'SATURDAY': 'Sábado',
-            'SUNDAY': 'Domingo'
-        }
-        self.number_days_dict = {
-            'Lunes': 0,
-            'Martes': 1,
-            'Miércoles': 2,
-            'Jueves': 3,
-            'Viernes': 4,
-            'Sábado': 5,
-            'Domingo': 6,
-        }
-        super(Helper, self).__init__()
-
-    def naive_to_datetime(self, nd):
-        if type(nd) == datetime:
-            if nd.tzinfo is not None and nd.tzinfo.utcoffset(nd) is not None:  # Is Aware
-                return nd
-            else:  # Is Naive
-                return self.tz.localize(nd)
-
-        elif type(nd) == date:
-            d = nd
-            t = time(0, 0)
-            new_date = datetime.combine(d, t)
-            return self.tz.localize(new_date)
-
-    def get_name_day(self, datetime_now):
-        name_day = date(datetime_now.year, datetime_now.month, datetime_now.day)
-        return self.days_dict[name_day.strftime('%A').upper()]
-
-    def get_number_day(self, dt):
-        return self.number_days_dict[self.get_name_day(dt)]
-
-    def start_datetime(self, back_days):
-        start_date = date.today() - timedelta(days=back_days)
-        return self.naive_to_datetime(start_date)
-
-    def end_datetime(self, back_days):
-        end_date = self.start_datetime(back_days) + timedelta(days=1)
-        return self.naive_to_datetime(end_date)
-
-    def parse_to_datetime(self, dt):
-        day = int(dt.split('-')[0])
-        month = int(dt.split('-')[1])
-        year = int(dt.split('-')[2])
-        parse_date = date(year, month, day)
-        return self.naive_to_datetime(parse_date)
-
-    def are_equal_lists(self, list_1, list_2):
-        """
-         Checks if two lists are identical
-        """
-        list_1 = self.items_list_to_int(list_1)
-        list_2 = self.items_list_to_int(list_2)
-
-        list_1.sort()
-        list_2.sort()
-
-        if len(list_1) != len(list_2):
-            return False
-        else:
-            for element in range(0, len(list_1)):
-                if list_1[element] != list_2[element]:
-                    return False
-
-        return True
-
-    @staticmethod
-    def get_week_number(dt):
-        return dt.isocalendar()[1]
-
-    @staticmethod
-    def items_list_to_int(list_to_cast):
-        """
-        Evaluates each of the elements of the list received and casts them to integers
-        """
-        cast_list = []
-        for item in range(0, len(list_to_cast)):
-            cast_list.append(int(list_to_cast[item]))
-
-        return cast_list
-
-
 class LeastSquares(object):
     def __init__(self, x: list, y: list):
         super(LeastSquares, self).__init__()
@@ -278,1032 +185,6 @@ class KitchenHelper(object):
         self.__all_processed_products = ProcessedProduct.objects. \
             select_related('ticket'). \
             all()
-
-
-class SalesHelper(object):
-    def __init__(self):
-        self.__all_tickets = None
-        self.__all_tickets_details = None
-        self.__all_extra_ingredients = None
-        super(SalesHelper, self).__init__()
-
-    def set_all_tickets(self):
-        self.__all_tickets = TicketBase.objects.select_related('seller').all()
-
-    def set_all_tickets_details(self):
-        self.__all_tickets_details = TicketDetail.objects. \
-            select_related('ticket'). \
-            select_related('cartridge'). \
-            select_related('ticket__seller'). \
-            select_related('package_cartridge'). \
-            all()
-
-    def set_all_extra_ingredients(self):
-        self.__all_extra_ingredients = TicketExtraIngredient.objects. \
-            select_related('ticket_detail'). \
-            select_related('extra_ingredient'). \
-            select_related('extra_ingredient__ingredient'). \
-            all()
-
-    def get_all_tickets(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_tickets is None:
-            self.set_all_tickets()
-        return self.__all_tickets
-
-    def get_all_tickets_details(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_tickets_details is None:
-            self.set_all_tickets_details()
-        return self.__all_tickets_details
-
-    def get_tickets_details(self, initial_date, final_date):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_tickets_details is None:
-            self.set_all_tickets_details()
-        return self.__all_tickets_details.filter(ticket__created_at__range=[initial_date, final_date])
-
-    def get_all_extra_ingredients(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_extra_ingredients is None:
-            self.set_all_extra_ingredients()
-        return self.__all_extra_ingredients
-
-    def get_years_list(self):
-        """
-        Returns a list of all the years in which there have been sales
-        """
-        years_list = []
-
-        for ticket in self.get_all_tickets():
-            if ticket.created_at.year not in years_list:
-                years_list.append(ticket.created_at.year)
-
-        return years_list
-
-    def get_tickets_today_list(self):
-        helper = Helper()
-
-        tickets_list = []
-        filtered_tickets = self.get_all_tickets().filter(created_at__gte=helper.naive_to_datetime(date.today()))
-
-        for ticket in filtered_tickets:
-            ticket_object = {
-                'ticket_parent': ticket,
-                'order_number': ticket.order_number,
-                'cartridges': [],
-                'packages': [],
-                'total': Decimal(0.00),
-                'is_active': ticket.is_active,
-            }
-
-            for ticket_details in self.get_all_tickets_details():
-                if ticket_details.ticket == ticket:
-                    if ticket_details.cartridge:
-                        cartridge_object = {
-                            'cartridge': ticket_details.cartridge,
-                            'quantity': ticket_details.quantity
-                        }
-                        ticket_object['cartridges'].append(cartridge_object)
-                        ticket_object['total'] += ticket_details.price
-                    elif ticket_details.package_cartridge:
-                        package_cartridge_object = {
-                            'package': ticket_details.package_cartridge,
-                            'quantity': ticket_details.quantity
-                        }
-                        ticket_object['packages'].append(package_cartridge_object)
-                        ticket_object['total'] += ticket_details.price
-
-            tickets_list.append(ticket_object)
-
-        return tickets_list
-
-    def get_dates_range_json(self):
-        """
-        Returns a JSON with a years list.
-        The years list contains years objects that contains a weeks list
-            and the Weeks list contains a weeks objects with two attributes:
-            start date and final date. Ranges of each week.
-        """
-        helper = Helper()
-        try:
-            min_year = self.get_all_tickets().aggregate(Min('created_at'))['created_at__min'].year
-            max_year = self.get_all_tickets().aggregate(Max('created_at'))['created_at__max'].year
-            years_list = []  # [2015:object, 2016:object, 2017:object, ...]
-        except:
-            min_year = datetime.now().year
-            max_year = datetime.now().year
-            years_list = []  # [2015:object, 2016:object, 2017:object, ...]
-
-        while max_year >= min_year:
-            year_object = {  # 2015:object or 2016:object or 2017:object ...
-                'year': max_year,
-                'weeks_list': [],
-            }
-
-            tickets_per_year = self.get_all_tickets().filter(
-                created_at__range=[helper.naive_to_datetime(date(max_year, 1, 1)),
-                                   helper.naive_to_datetime(date(max_year, 12, 31))])
-            for ticket_item in tickets_per_year:
-                if len(year_object['weeks_list']) == 0:
-                    """
-                    Creates a new week_object in the weeks_list of the actual year_object
-                    """
-                    week_object = {
-                        'week_number': ticket_item.created_at.isocalendar()[1],
-                        'start_date': ticket_item.created_at.date().strftime("%d-%m-%Y"),
-                        'end_date': ticket_item.created_at.date().strftime("%d-%m-%Y"),
-                    }
-                    year_object['weeks_list'].append(week_object)
-                    # End if
-                else:
-                    """
-                    Validates if exists some week with an similar week_number of the actual year
-                    If exists a same week in the list validates the start_date and the end_date,
-                    In each case valid if there is an older start date or a more current end date
-                        if it is the case, update the values.
-                    Else creates a new week_object with the required week number
-                    """
-                    existing_week = False
-                    for week_object in year_object['weeks_list']:
-
-                        if week_object['week_number'] == ticket_item.created_at.isocalendar()[1]:
-                            # There's a same week number
-                            if datetime.strptime(week_object['start_date'],
-                                                 "%d-%m-%Y").date() > ticket_item.created_at.date():
-                                week_object['start_date'] = ticket_item.created_at.date().strftime("%d-%m-%Y")
-                            elif datetime.strptime(week_object['end_date'],
-                                                   "%d-%m-%Y").date() < ticket_item.created_at.date():
-                                week_object['end_date'] = ticket_item.created_at.date().strftime("%d-%m-%Y")
-
-                            existing_week = True
-                            break
-
-                    if not existing_week:
-                        # There's a different week number
-                        week_object = {
-                            'week_number': ticket_item.created_at.isocalendar()[1],
-                            'start_date': ticket_item.created_at.date().strftime("%d-%m-%Y"),
-                            'end_date': ticket_item.created_at.date().strftime("%d-%m-%Y"),
-                        }
-                        year_object['weeks_list'].append(week_object)
-
-                        # End else
-            years_list.append(year_object)
-            max_year -= 1
-        # End while
-        return json.dumps(years_list)
-
-    def get_sales_list(self, start_dt, final_dt):
-        """
-        Gets the following properties for each week's day: Name, Date and Earnings
-        """
-        helper = Helper()
-        limit_day = start_dt + timedelta(days=1)
-        total_days = (final_dt - start_dt).days
-        week_sales_list = []
-        count = 1
-        total_earnings = 0
-
-        while count <= total_days:
-            day_tickets = self.get_all_tickets().filter(created_at__range=[start_dt, limit_day])
-            day_object = {
-                'date': str(start_dt.date().strftime('%d-%m-%Y')),
-                'day_name': None,
-                'earnings': None,
-                'number_day': helper.get_number_day(start_dt),
-            }
-
-            for ticket_item in day_tickets:
-                for ticket_detail_item in self.get_all_tickets_details():
-                    if ticket_detail_item.ticket == ticket_item:
-                        total_earnings += ticket_detail_item.price
-
-            day_object['day_name'] = helper.get_name_day(start_dt.date())
-            day_object['earnings'] = str(total_earnings)
-
-            week_sales_list.append(day_object)
-
-            # Reset data
-            limit_day += timedelta(days=1)
-            start_dt += timedelta(days=1)
-            total_earnings = 0
-            count += 1
-
-        return week_sales_list
-
-    def get_sales_actual_week(self):
-        """
-        Gets the following properties for each week's day: Name, Date and Earnings
-        """
-        helper = Helper()
-        week_sales_list = []
-        total_earnings = 0
-        days_to_count = helper.get_number_day(datetime.now())
-        day_limit = days_to_count
-        start_date_number = 0
-
-        while start_date_number <= day_limit:
-            day_object = {
-                'date': str(helper.start_datetime(days_to_count).date().strftime('%d-%m-%Y')),
-                'day_name': None,
-                'earnings': None,
-                'number_day': helper.get_number_day(helper.start_datetime(days_to_count).date()),
-            }
-
-            day_tickets = self.get_all_tickets().filter(
-                created_at__range=[helper.start_datetime(days_to_count), helper.end_datetime(days_to_count)])
-
-            for ticket_item in day_tickets:
-                for ticket_detail_item in self.get_all_tickets_details():
-                    if ticket_detail_item.ticket == ticket_item:
-                        total_earnings += ticket_detail_item.price
-
-            day_object['earnings'] = str(total_earnings)
-            day_object['day_name'] = helper.get_name_day(helper.start_datetime(days_to_count).date())
-
-            week_sales_list.append(day_object)
-
-            # restarting counters
-            days_to_count -= 1
-            total_earnings = 0
-            start_date_number += 1
-
-        return json.dumps(week_sales_list)
-
-    def get_tickets_list(self, initial_date, final_date):
-        """
-        :rtype: list
-        :param initial_date: datetime
-        :param final_date: datetime
-        """
-        all_tickets = self.get_all_tickets().filter(
-            created_at__range=(initial_date, final_date)).order_by('-created_at')
-        all_tickets_details = self.get_all_tickets_details()
-        tickets_list = []
-
-        for ticket in all_tickets:
-            ticket_object = {
-                'id': ticket.id,
-                'order_number': ticket.order_number,
-                'created_at': datetime.strftime(ticket.created_at, "%B %d, %Y, %H:%M:%S %p"),
-                'seller': ticket.seller.username,
-                'ticket_details': {
-                    'cartridges': [],
-                    'packages': [],
-                },
-                'total': 0,
-            }
-
-            for ticket_detail in all_tickets_details:
-                if ticket_detail.ticket == ticket:
-                    ticket_detail_object = {}
-                    if ticket_detail.cartridge:
-                        ticket_detail_object = {
-                            'name': ticket_detail.cartridge.name,
-                            'quantity': ticket_detail.quantity,
-                            'price': float(ticket_detail.price),
-                        }
-                        ticket_object['ticket_details']['cartridges'].append(ticket_detail_object)
-                    elif ticket_detail.package_cartridge:
-                        ticket_detail_object = {
-                            'name': ticket_detail.package_cartridge.name,
-                            'quantity': ticket_detail.quantity,
-                            'price': float(ticket_detail.price),
-                        }
-                        ticket_object['ticket_details']['packages'].append(ticket_detail_object)
-
-                    ticket_object['total'] += float(ticket_detail.price)
-
-                    try:
-                        ticket_object['ticket_details'].append(ticket_detail_object)
-                    except Exception as e:
-                        pass
-            ticket_object['total'] = str(ticket_object['total'])
-            tickets_list.append(ticket_object)
-        return tickets_list
-
-    def get_new_order_number(self):
-        if self.__all_tickets is None:
-            self.set_all_tickets()
-
-        order_numbers_list = []
-        for ticket in self.get_all_tickets():
-            order_numbers_list.append(ticket.order_number)
-
-        try:
-            return max(order_numbers_list) +1
-
-        except ValueError:
-            return 1
-
-
-class ProductsHelper(object):
-    def __init__(self):
-        super(ProductsHelper, self).__init__()
-        self.__all_cartridges = None
-        self.__all_packages_cartridges = None
-        self.__all_supplies = None
-        self.__all_extra_ingredients = None
-        self.__all_cartridges_recipes = None
-        self.__all_tickets_details = None
-        self.__elements_in_warehouse = None
-        self.__predictions = None
-        self.__required_supplies_list = None
-        self.__today_popular_cartridge = None
-        self.__all_packages_cartridges_recipes = None
-        self.__always_popular_cartridge = None
-
-    def get_all_supplies(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_supplies is None:
-            self.set_all_supplies()
-        return self.__all_supplies
-
-    @property
-    def all_cartridges(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_cartridges is None:
-            self.set_all_cartridges()
-        return self.__all_cartridges
-
-    def get_all_packages_cartridges(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_packages_cartridges is None:
-            self.set_all_packages_cartridges()
-        return self.__all_packages_cartridges
-
-    def get_all_extra_ingredients(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_extra_ingredients is None:
-            self.set_all_extra_ingredients()
-
-        return self.__all_extra_ingredients
-
-    def get_all_cartridges_recipes(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_cartridges_recipes is None:
-            self.set_all_cartridges_recipes()
-
-        return self.__all_cartridges_recipes
-
-    def get_all_packages_cartridges_recipes(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_packages_cartridges_recipes is None:
-            self.set_all_package_cartridges_recipes()
-
-        return self.__all_packages_cartridges_recipes
-
-    def get_all_elements_in_warehouse(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__elements_in_warehouse is None:
-            self.set_elements_in_warehouse()
-        return self.__elements_in_warehouse
-
-    def get_required_supplies(self):
-        """
-        :rtype: list
-        """
-        required_supplies_list = []
-        all_cartridges = self.all_cartridges()
-        predictions = self.get_predictions_supplies()
-        supplies_on_stock = self.get_all_elements_in_warehouse().filter(status="ST")
-
-        ingredients = self.get_all_cartridges_recipes()
-
-        for prediction in predictions:
-            for cartridge in all_cartridges:
-                if prediction['cartridge'] == cartridge:
-
-                    ingredientes = ingredients.filter(cartridge=cartridge)
-
-                    for ingredient in ingredientes:
-
-                        supply = ingredient.supply
-                        name = ingredient.supply.name
-                        cost = ingredient.supply.presentation_cost
-                        measurement_unit = ingredient.supply.unit_conversion(ingredient.quantity)
-                        measurement_quantity = ingredient.supply.measurement_conversion(ingredient.quantity)
-                        supplier_unit = ingredient.supply.unit_conversion(ingredient.supply.measurement_quantity)
-                        supplier_quantity = ingredient.supply.measurement_conversion(
-                            ingredient.supply.measurement_quantity)
-                        quantity = ingredient.quantity
-                        supplier = ingredient.supply.supplier
-
-                        count = 0
-
-                        required_supply_object = {
-                            'supply': supply,
-                            'name': name,
-                            'cost': cost,
-                            'measurement_unit': measurement_unit,
-                            'measurement_quantity': measurement_quantity,
-                            'supplier_unit': supplier_unit,
-                            'supplier_quantity': supplier_quantity,
-                            'quantity': quantity,
-                            'supplier': supplier,
-                            'stock': 0,
-                            'required': 0,
-                            'full_cost': 0,
-                        }
-
-                        if len(required_supplies_list) == 0:
-                            count = 1
-                        else:
-                            for required_supplies in required_supplies_list:
-                                if required_supplies['name'] == name:
-                                    required_supplies['quantity'] += quantity
-                                    count = 0
-                                    break
-                                else:
-                                    count = 1
-                        if count == 1:
-                            required_supplies_list.append(required_supply_object)
-
-        for required_supply in required_supplies_list:
-            if len(supplies_on_stock) > 0:
-                for supply_on_stock in supplies_on_stock:
-                    if supply_on_stock.supply == required_supply['supply']:
-                        required_supply['stock'] = supply_on_stock.quantity
-                        required_supply['required'] = max(0, required_supply['quantity'] - required_supply['stock'])
-                        required_supply['full_cost'] = \
-                            required_supply['cost'] * \
-                            math.ceil(required_supply['required'] / required_supply['measurement_quantity'])
-                        break
-                    else:
-                        required_supply['required'] = max(0, required_supply['quantity'] - required_supply['stock'])
-                        required_supply['full_cost'] = \
-                            required_supply['cost'] * \
-                            math.ceil(required_supply['required'] / required_supply['measurement_quantity'])
-                        required_supply['full_cost'] = \
-                            required_supply['cost'] * \
-                            math.ceil(required_supply['required'] / required_supply['measurement_quantity'])
-
-            else:
-                required_supply['required'] = max(0, required_supply['quantity'] - required_supply['stock'])
-                required_supply['full_cost'] = \
-                    required_supply['cost'] * \
-                    math.ceil(required_supply['required'] / required_supply['measurement_quantity'])
-
-        return required_supplies_list
-
-    def get_always_popular_cartridge(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__always_popular_cartridge is None:
-            self.set_always_popular_cartridge()
-        return self.__always_popular_cartridge
-
-    def get_predictions_supplies(self):
-        """ :rtype: list """
-        if self.__predictions is None:
-            self.set_predictions()
-        return self.__predictions
-
-    def get_supplies_on_stock_list(self):
-        """ :rtype: list """
-        stock_list = []
-        all_elements = self.__elements_in_warehouse.filter(status='ST')
-        if all_elements.count() > 0:
-            for element in all_elements:
-                stock_object = {
-                    'name': element.supply.name,
-                    'quantity': element.quantity,
-                }
-                stock_list.append(stock_object)
-
-        return stock_list
-
-    def get_today_popular_cartridge(self):
-        if self.__today_popular_cartridge is None:
-            self.set_today_popular_cartridge()
-        return self.__always_popular_cartridge
-
-    def set_all_supplies(self):
-        self.__all_supplies = Supply.objects. \
-            select_related('category'). \
-            select_related('supplier'). \
-            select_related('location').order_by('name')
-
-    def set_all_cartridges(self):
-        self.__all_cartridges = Cartridge.objects.order_by('subcategory', 'name')
-
-    def set_all_packages_cartridges(self):
-        self.__all_packages_cartridges = PackageCartridge.objects.all()
-
-    def set_all_cartridges_recipes(self):
-        self.__all_cartridges_recipes = CartridgeRecipe.objects. \
-            select_related('cartridge'). \
-            select_related('supply'). \
-            all()
-
-    def set_all_package_cartridges_recipes(self):
-        self.__all_packages_cartridges_recipes = PackageCartridgeRecipe.objects. \
-            select_related('package_cartridge'). \
-            select_related('cartridge'). \
-            all()
-
-    def set_all_extra_ingredients(self):
-        self.__all_extra_ingredients = ExtraIngredient.objects. \
-            select_related('ingredient'). \
-            select_related('cartridge'). \
-            all()
-
-    def set_predictions(self):
-        sales_helper = SalesHelper()
-        all_tickets_details = sales_helper.get_all_tickets_details()
-
-        prediction_list = []
-
-        for ticket_details in all_tickets_details:
-            cartridge_object = {
-                'cartridge': ticket_details.cartridge,
-                'cantidad': 1,
-            }
-
-            prediction_list.append(cartridge_object)
-
-        self.__predictions = prediction_list
-
-    def set_all_tickets_details(self):
-        self.__all_tickets_details = TicketDetail.objects.select_related(
-            'ticket').select_related('cartridge').select_related('package_cartridge').all()
-
-    def set_always_popular_cartridge(self):
-        sales_helper = SalesHelper()
-        cartridges_frequency_dict = {}
-        for cartridge in self.all_cartridges():
-            cartridges_frequency_dict[cartridge.id] = {
-                'frequency': 0,
-                'name': cartridge.name,
-            }
-        for ticket_detail in sales_helper.get_all_tickets_details():
-            if ticket_detail.cartridge:
-                ticket_detail_id = ticket_detail.cartridge.id
-                ticket_detail_frequency = ticket_detail.quantity
-                cartridges_frequency_dict[ticket_detail_id]['frequency'] += ticket_detail_frequency
-
-        for element in cartridges_frequency_dict:
-            if self.__always_popular_cartridge is None:
-                """ Base case """
-                self.__always_popular_cartridge = {
-                    'id': element,
-                    'name': cartridges_frequency_dict[element]['name'],
-                    'frequency': cartridges_frequency_dict[element]['frequency'],
-                }
-            else:
-                if cartridges_frequency_dict[element]['frequency'] > self.__always_popular_cartridge['frequency']:
-                    self.__always_popular_cartridge = {
-                        'id': element,
-                        'name': cartridges_frequency_dict[element]['name'],
-                        'frequency': cartridges_frequency_dict[element]['frequency'],
-                    }
-
-    def set_elements_in_warehouse(self):
-        self.__elements_in_warehouse = Warehouse.objects.select_related('supply').all().order_by('supply__name')
-
-    def set_today_popular_cartridge(self):
-        sales_helper = SalesHelper()
-        cartridges_frequency_dict = {}
-        helper = Helper()
-        start_date = helper.naive_to_datetime(date.today())
-        limit_day = helper.naive_to_datetime(start_date + timedelta(days=1))
-        filtered_ticket_details = sales_helper.get_tickets_details(start_date, limit_day)
-
-        for cartridge in self.all_cartridges():
-            cartridges_frequency_dict[cartridge.id] = {
-                'frequency': 0,
-                'name': cartridge.name,
-            }
-
-        for ticket_detail in filtered_ticket_details:
-            if ticket_detail.cartridge:
-                ticket_detail_id = ticket_detail.cartridge.id
-                ticket_detail_frequency = ticket_detail.quantity
-                cartridges_frequency_dict[ticket_detail_id]['frequency'] += ticket_detail_frequency
-
-        for element in cartridges_frequency_dict:
-            if self.__today_popular_cartridge is None:
-                """ Base case """
-                self.__today_popular_cartridge = {
-                    'id': element,
-                    'name': cartridges_frequency_dict[element]['name'],
-                    'frequency': cartridges_frequency_dict[element]['frequency'],
-                }
-            else:
-                if cartridges_frequency_dict[element]['frequency'] > self.__today_popular_cartridge['frequency']:
-                    self.__today_popular_cartridge = {
-                        'id': element,
-                        'name': cartridges_frequency_dict[element]['name'],
-                        'frequency': cartridges_frequency_dict[element]['frequency'],
-                    }
-
-
-class DinersHelper(object):
-    def __init__(self):
-        self.__all_diners = None
-        self.__all_access_logs = None
-        super(DinersHelper, self).__init__()
-
-    def get_all_diners_logs_list(self, initial_date, final_date):
-        helper = Helper()
-        diners_logs_list = []
-
-        diners_logs_objects = self.get_access_logs(initial_date, final_date)
-
-        for diner_log in diners_logs_objects:
-            diner_log_object = {
-                'rfid': diner_log.RFID,
-                'access': datetime.strftime(timezone.localtime(diner_log.access_to_room), "%B %d, %I, %H:%M:%S %p"),
-                'number_day': helper.get_number_day(diner_log.access_to_room),
-            }
-            if diner_log.diner:
-                diner_log_object['SAP'] = diner_log.diner.employee_number
-                diner_log_object['name'] = diner_log.diner.name
-            else:
-                diner_log_object['SAP'] = ''
-                diner_log_object['name'] = ''
-            diners_logs_list.append(diner_log_object)
-        return diners_logs_list
-
-    def get_weeks_entries(self, initial_dt, final_dt):
-        """
-        Gets the following properties for each week's day: Name, Date and Earnings
-        """
-        if self.__all_access_logs is None:
-            self.set_all_access_logs()
-
-        helper = Helper()
-        limit_day = initial_dt + timedelta(days=1)
-        weeks_list = []
-        count = 1
-        total_days = (final_dt - initial_dt).days
-
-        while count <= total_days:
-            diners_entries = self.__all_access_logs.filter(access_to_room__range=[initial_dt, limit_day])
-            day_object = {
-                'date': str(timezone.localtime(initial_dt).date().strftime('%d-%m-%Y')),
-                'day_name': helper.get_name_day(initial_dt.date()), 'entries': diners_entries.count(),
-                'number_day': helper.get_number_day(initial_dt)}
-
-            weeks_list.append(day_object)
-
-            # Reset data
-            limit_day += timedelta(days=1)
-            initial_dt += timedelta(days=1)
-            count += 1
-
-        return weeks_list
-
-    def get_access_logs(self, initial_date, final_date):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_access_logs is None:
-            self.set_all_access_logs()
-
-        return self.__all_access_logs. \
-            filter(access_to_room__range=(initial_date, final_date)). \
-            order_by('-access_to_room')
-
-    def get_access_logs_today(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_access_logs is None:
-            self.set_all_access_logs()
-
-        helper = Helper()
-        year = int(datetime.now().year)
-        month = int(datetime.now().month)
-        day = int(datetime.now().day)
-        initial_date = helper.naive_to_datetime(date(year, month, day))
-        final_date = helper.naive_to_datetime(initial_date + timedelta(days=1))
-        return self.__all_access_logs. \
-            filter(access_to_room__range=(initial_date, final_date)). \
-            order_by('-access_to_room')
-
-    def get_all_access_logs(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_access_logs is None:
-            self.set_all_access_logs()
-        return self.__all_access_logs
-
-    def get_diners_per_hour_json(self):
-        hours_list = []
-        hours_to_count = 12
-        start_hour = 5
-        customer_count = 0
-        logs = self.get_access_logs_today()
-
-        while start_hour <= hours_to_count:
-            hour = {'count': None, }
-            for log in logs:
-                log_datetime = str(log.access_to_room)
-                log_date, log_time = log_datetime.split(" ")
-
-                if log_time.startswith("0" + str(start_hour)):
-                    customer_count += 1
-                hour['count'] = customer_count
-
-            hours_list.append(hour)
-            customer_count = 0
-            start_hour += 1
-
-        return json.dumps(hours_list)
-
-    def get_diners_actual_week(self):
-        if self.__all_access_logs is None:
-            self.set_all_access_logs()
-        helper = Helper()
-        week_diners_list = []
-        total_entries = 0
-        days_to_count = helper.get_number_day(date.today())
-        day_limit = days_to_count
-        start_date_number = 0
-
-        while start_date_number <= day_limit:
-            day_object = {
-                'date': str(helper.start_datetime(days_to_count).date().strftime('%d-%m-%Y')),
-                'day_name': None,
-                'entries': None,
-                'number_day': helper.get_number_day(helper.start_datetime(days_to_count).date())
-            }
-
-            logs = self.__all_access_logs. \
-                filter(access_to_room__range=[helper.start_datetime(days_to_count), helper.end_datetime(days_to_count)])
-
-            for _ in logs:
-                total_entries += 1
-
-            day_object['entries'] = str(total_entries)
-            day_object['day_name'] = helper.get_name_day(helper.start_datetime(days_to_count).date())
-
-            week_diners_list.append(day_object)
-
-            # restarting counters
-            days_to_count -= 1
-            total_entries = 0
-            start_date_number += 1
-
-        return json.dumps(week_diners_list)
-
-    def get_all_diners(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_diners is None:
-            self.set_all_diners()
-        return self.__all_diners
-
-    def set_all_access_logs(self):
-        self.__all_access_logs = AccessLog.objects.select_related('diner').order_by('-access_to_room')
-
-    def set_all_diners(self):
-        self.__all_diners = Diner.objects.all()
-
-
-class RatesHelper(object):
-    def __init__(self):
-        super(RatesHelper, self).__init__()
-        self.__all_satisfaction_ratings = None
-        self.__elements_to_evaluate = None
-
-    @property
-    def satisfaction_ratings(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__all_satisfaction_ratings is None:
-            self.set_all_satisfaction_ratings()
-        return self.__all_satisfaction_ratings
-
-    @property
-    def elements_to_evaluate(self):
-        """
-        :rtype: django.db.models.query.QuerySet
-        """
-        if self.__elements_to_evaluate is None:
-            self.set_elements_to_evaluate()
-        return self.__elements_to_evaluate
-
-    def get_dates_range(self):
-        """
-        Returns a JSON with a years list.
-        The years list contains years objects that contains a weeks list
-            and the Weeks list contains a weeks objects with two attributes:
-            start date and final date. Ranges of each week.
-        """
-        helper = Helper()
-        try:
-            min_year = self.satisfaction_ratings.aggregate(Min('creation_date'))['creation_date__min'].year
-            max_year = self.satisfaction_ratings.aggregate(Max('creation_date'))['creation_date__max'].year
-            years_list = []  # [2015:object, 2016:object, 2017:object, ...]
-        except:
-            min_year = datetime.now().year
-            max_year = datetime.now().year
-            years_list = []  # [2015:object, 2016:object, 2017:object, ...]
-
-        while max_year >= min_year:
-            year_object = {  # 2015:object or 2016:object or 2017:object ...
-                'year': max_year,
-                'weeks_list': [],
-            }
-
-            ratings_per_year = self.satisfaction_ratings.filter(
-                creation_date__range=[
-                    helper.naive_to_datetime(date(max_year, 1, 1)),
-                    helper.naive_to_datetime(date(max_year, 12, 31))])
-
-            for rating in ratings_per_year:
-                if not year_object['weeks_list']:
-                    """
-                    Creates a new week_object in the weeks_list of the actual year_object
-                    """
-                    week_object = {
-                        'week_number': rating.creation_date.isocalendar()[1],
-                        'start_date': rating.creation_date.date().strftime("%d-%m-%Y"),
-                        'end_date': rating.creation_date.date().strftime("%d-%m-%Y"),
-                    }
-                    year_object['weeks_list'].append(week_object)
-                    # End if
-                else:
-                    """
-                    Validates if exists some week with an similar week_number of the actual year
-                    If exists a same week in the list validates the start_date and the end_date,
-                    In each case valid if there is an older start date or a more current end date
-                        if it is the case, update the values.
-                    Else creates a new week_object with the required week number
-                    """
-                    existing_week = False
-                    for week_object in year_object['weeks_list']:
-
-                        if week_object['week_number'] == rating.creation_date.isocalendar()[1]:
-                            # There's a same week number
-                            if datetime.strptime(week_object['start_date'], "%d-%m-%Y").date() > \
-                                    rating.creation_date.date():
-                                week_object['start_date'] = rating.creation_date.date().strftime("%d-%m-%Y")
-                            elif datetime.strptime(week_object['end_date'], "%d-%m-%Y").date() < \
-                                    rating.creation_date.date():
-                                week_object['end_date'] = rating.creation_date.date().strftime("%d-%m-%Y")
-
-                            existing_week = True
-                            break
-
-                    if not existing_week:
-                        # There's a different week number
-                        week_object = {
-                            'week_number': rating.creation_date.isocalendar()[1],
-                            'start_date': rating.creation_date.date().strftime("%d-%m-%Y"),
-                            'end_date': rating.creation_date.date().strftime("%d-%m-%Y"),
-                        }
-                        year_object['weeks_list'].append(week_object)
-
-                        # End else
-            # End While
-            year_object['weeks_list'].reverse()
-            years_list.append(year_object)
-            max_year -= 1
-        # End while
-        return json.dumps(years_list)
-
-    def get_satisfaction_ratings(self, initial_date: datetime, final_date: datetime):
-        helper = Helper()
-        initial_date = helper.naive_to_datetime(initial_date)
-        final_date = helper.naive_to_datetime(final_date)
-        if self.__all_satisfaction_ratings is None:
-            self.set_all_satisfaction_ratings()
-        return self.__all_satisfaction_ratings.filter(
-            creation_date__range=[initial_date, final_date]).order_by('-creation_date')
-
-    def get_info_rates_list(self, initial_date: datetime, final_date: datetime):
-        """
-        Returns a list with all the rates data for te selected range
-        :rtype: list
-        """
-        helper = Helper()
-        week_suggestions_list = []
-
-        while initial_date <= final_date:
-            day_object = {
-                'date': str(initial_date.strftime('%d-%m-%Y')),
-                'day_name': None,
-                'total_rates': None,
-                'number_day': helper.get_number_day(initial_date),
-            }
-
-            filtered_suggestions = self.satisfaction_ratings.filter(
-                creation_date__range=[
-                    helper.naive_to_datetime(initial_date),
-                    helper.naive_to_datetime(initial_date + timedelta(days=1))])
-
-            day_object['total_rates'] = str(filtered_suggestions.count())
-            day_object['day_name'] = helper.get_name_day(initial_date)
-            week_suggestions_list.append(day_object)
-
-            # restarting counters
-            initial_date = initial_date + timedelta(days=1)
-
-        return week_suggestions_list
-
-    def get_info_rates_actual_week(self):
-        """
-        Gets the following properties for each week's day: Day name, Date, number day and total rates
-        :rtype: list
-        """
-        helper = Helper()
-        week_suggestions_list = []
-        days_to_count = helper.get_number_day(datetime.now())
-        day_limit = days_to_count
-        start_date_number = 0
-
-        while start_date_number <= day_limit:
-            day_object = {
-                'date': str(helper.start_datetime(days_to_count).date().strftime('%d-%m-%Y')),
-                'day_name': None,
-                'total_rates': None,
-                'number_day': helper.get_number_day(helper.start_datetime(days_to_count).date()),
-            }
-
-            filtered_suggestions = self.satisfaction_ratings.filter(
-                creation_date__range=[helper.start_datetime(days_to_count), helper.end_datetime(days_to_count)])
-
-            day_object['total_rates'] = str(filtered_suggestions.count())
-            day_object['day_name'] = helper.get_name_day(helper.start_datetime(days_to_count).date())
-
-            week_suggestions_list.append(day_object)
-
-            # restarting counters
-            days_to_count -= 1
-            start_date_number += 1
-
-        return json.dumps(week_suggestions_list)
-
-    def get_info_suggestions_actual_week(self):
-        """
-        Gets the following properties for each week's day: Day name, Date, number day and total suggestions
-        :rtype: list
-        """
-        helper = Helper()
-        week_suggestions_list = []
-        days_to_count = helper.get_number_day(datetime.now())
-        day_limit = days_to_count
-        start_date_number = 0
-
-        while start_date_number <= day_limit:
-            day_object = {
-                'date': str(helper.start_datetime(days_to_count).date().strftime('%d-%m-%Y')),
-                'day_name': None,
-                'total_suggestions': None,
-                'number_day': helper.get_number_day(helper.start_datetime(days_to_count).date()),
-            }
-
-            filtered_suggestions = self.satisfaction_ratings.filter(
-                creation_date__range=[helper.start_datetime(days_to_count), helper.end_datetime(days_to_count)]). \
-                exclude(suggestion__isnull=True)
-
-            day_object['total_suggestions'] = str(filtered_suggestions.count())
-            day_object['day_name'] = helper.get_name_day(helper.start_datetime(days_to_count).date())
-
-            week_suggestions_list.append(day_object)
-
-            # restarting counters
-            days_to_count -= 1
-            start_date_number += 1
-
-        return json.dumps(week_suggestions_list)
-
-    def set_all_satisfaction_ratings(self):
-        self.__all_satisfaction_ratings = SatisfactionRating.objects \
-            .prefetch_related('elements').all()
-
-    def set_elements_to_evaluate(self):
-        self.__elements_to_evaluate = ElementToEvaluate.objects.order_by('priority').all()
 
 
 class CartHelper(object):
