+++ sales/templates/sales/sales.html
@@ -2,12 +2,6 @@
 
 {% load static %}
 
-{% block link %}
-  <!--
-  <link rel="stylesheet" href="{% static 'css/sweetalert.css' %}">
-  -->
-{% endblock link %}
-
 {% block page_title %}
   {{ page_title }}
 {% endblock %}
@@ -18,6 +12,7 @@
     <div class="col-xs-12">
       <div class="text-xs-right mb-2">
         <form class="form-inline" id="dates-range-form">
+          <label for="dt-year" hidden></label>
           <select class="custom-select" id="dt-year">
           </select>
           <select class="custom-select" id="dt-week" >
@@ -101,7 +96,7 @@
                 </span>
                 {% endfor %}
               </td>
-              <td>{{ ticket.ticket_parent.seller }}</td>
+              <td>{{ ticket.ticket_parent.cashier }}</td>
               <td class="td-total">
                 {{ ticket.total }}
               </td>
@@ -192,7 +187,6 @@
 
     var sales_week = JSON.parse('{{ sales_week | safe }}');
     var dates_range = JSON.parse('{{ dates_range | safe }}');
-
     /**
      * returns a list with the earnings of the week per day
      */
@@ -667,7 +661,7 @@
               "<td class='header-date'>" + tickets_objects[i].created_at + "</td>" +
               "<td class='header-products'>" + cartridges_list + "</td>" +
               "<td class='header-packages'>" + packages_list + "</td>" +
-              "<td>" + tickets_objects[i].seller + "</td>" +
+              "<td>" + tickets_objects[i].cashier + "</td>" +
               "<td class='td-total'>" + tickets_objects[i].total + "</td>" +
               "<td class='header-actions'>" +
               "<span class='sales-actions delete-ticket'><i class='material-icons text-muted'>delete</i></span>" +
@@ -861,7 +855,7 @@
 
       function delete_ticket() {
         $.ajax({
-          url: "{% url 'sales:delete-sale' %}",
+          url: "{% url 'sales:sales_delete' %}",
           type: 'POST',
           data: {
             csrfmiddlewaretoken: '{{ csrf_token }}',
