+++ sales/urls.py
@@ -8,14 +8,11 @@ app_name = 'sales'
 urlpatterns = [
     # sales
     url(r'^sales/$', views.sales, name='sales'),
-    url(r'^sales/new/$', views.new_sale, name='new_sale'),
-    url(r'^sales/delete/$', views.delete_sale, name='delete-sale'),
+    url(r'^sales/new/$', views.new_sale, name='sales_new'),
+    url(r'^sales/delete/$', views.delete_sale, name='sales_delete'),
 
 ]
 
 if settings.DEBUG:
     # NEW SALES UPGRADE
-    urlpatterns.append(url('^test/sales/$', views.test_sales, name='test_sales'))
     urlpatterns.append(url('^test/sales/update/$', views.test_sales_update, name='test_sales_update'))
-    urlpatterns.append(url('^test/sales/new/$', views.test_sales_new, name='test_sales_new'))
-    urlpatterns.append(url('^test/sales/delete/$', views.test_sales_delete, name='test_sales_delete'))
